pyproject.toml
[project]
name = "quantum-computing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "qiskit-aer>=0.17.1",
    "qiskit-machine-learning[torch]>=0.8.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "yfinance>=0.2.65",
]
